% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bks.R
\name{append.bks.to.instance.frame}
\alias{append.bks.to.instance.frame}
\title{Extend an Instance Data Frame by Adding the Best-Known Solutions from
  Literature}
\usage{
append.bks.to.instance.frame(instances, results,
  are.objective.values.ints = TRUE)
}
\arguments{
\item{instances}{the instance frame}

\item{results}{the results data frame}

\item{are.objective.values.ints}{are the objective values integers?}
}
\value{
the expanded instance frame
}
\description{
Take an results data frame and extract, for each instance, the
  best-known solutions as well as the accompanying references.
}
