% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_column.R
\name{create.column}
\alias{create.column}
\title{Create a Column Descriptor}
\usage{
create.column(title, description, type)
}
\arguments{
\item{title}{the column title}

\item{description}{the column description}

\item{type}{the column type}
}
\value{
the column description
}
\description{
Create a column descriptor as used by our code generation
}
