% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_column.R
\name{create.columns}
\alias{create.columns}
\title{Create a Set of Column Descriptors along with Auto-Generated Code}
\usage{
create.columns(columns, conditions = character(0),
  mergers = character(0))
}
\arguments{
\item{columns}{the column descriptors}

\item{conditions}{the auto-generated validation code}

\item{mergers}{the auto-generated merger code}
}
\value{
a set of column descriptors
}
\description{
Well, create such a set.
}
