% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_standard_meta_columns.R
\name{create.standard.meta.columns}
\alias{create.standard.meta.columns}
\title{Create the Standard Meta-Data Column Set}
\usage{
create.standard.meta.columns(are.objective.values.ints = TRUE,
  objective.value.lower.bound = 0L,
  objective.value.upper.bound = NA_integer_)
}
\arguments{
\item{are.objective.values.ints}{are objective values integers?}

\item{objective.value.lower.bound}{a hard lower bound for objective values}

\item{objective.value.upper.bound}{a hard upper bound for objective values}
}
\value{
the set of standard meta-data columns
}
\description{
Create the set of standard meta-data columns
}
