% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_standard_result_columns.R
\name{create.standard.result.columns}
\alias{create.standard.result.columns}
\title{Create the Standard Result Column Set}
\usage{
create.standard.result.columns(are.objective.values.ints = TRUE,
  is.time.int = TRUE, objective.value.lower.bound = 0L,
  objective.value.upper.bound = NA_integer_)
}
\arguments{
\item{are.objective.values.ints}{are objective values integers?}

\item{is.time.int}{are time values integers?}

\item{objective.value.lower.bound}{a hard lower bound for objective values}

\item{objective.value.upper.bound}{a hard upper bound for objective values}
}
\value{
the set of standard result columns
}
\description{
Create the set of standard result (statistic) columns
}
