% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_stat_columns.R
\name{create.stat.columns}
\alias{create.stat.columns}
\title{Create Names and Conditions for Statistics Columns}
\usage{
create.stat.columns(title, description, number.column = "n.runs",
  is.col.integer = FALSE, lower.bound = 0L,
  upper.bound = NA_integer_)
}
\arguments{
\item{title}{the variable name}

\item{description}{the variable description}

\item{number.column}{the name of the column with the number of repetitions}

\item{is.col.integer}{is the variable/column integer valued?}

\item{lower.bound}{an optional lower bound}

\item{upper.bound}{an optional upper bound}
}
\value{
a list with the auto-generated members and code
}
\description{
We can store several types of statistics from an experiment
  measuring one variable. These include minimum, mean, median, mode, and
  maximum as well as standard deviation. There are several sanity constraints
  imposed on them, e.g., the minimum cannot be bigger than the maximum and if
  the standard deviation is 0 they must be the same, etc. Here we try to
  generate "column names" and conditions for a given variable.
}
