% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_result_data_frame_functions.R
\name{generate.loader.functions}
\alias{generate.loader.functions}
\title{Generate All Code Necessary for Fully Loading, Expanding, and Validating a Set of Results}
\usage{
generate.loader.functions(meta.columns = create.standard.meta.columns(),
  result.columns = create.standard.result.columns(),
  function.name = .func.result.load.all,
  validate.function.name = .func.result.data.frame.validate,
  expand.function.name = .func.result.data.frame.expand,
  frame.load.function.name = .func.result.data.frame.load)
}
\arguments{
\item{meta.columns}{the meta-data columns. Their value can either be
specified as parameter of the generated function or inside the result csv
file.}

\item{result.columns}{the values of these columns can only be specified in
the csv file}

\item{function.name}{the function name to use for the generated function}

\item{validate.function.name}{the name of the function generated by
\code{generate.result.data.frame.validate.function}}

\item{expand.function.name}{the name of the function generated by
\code{generate.result.data.frame.expand.function}}

\item{frame.load.function.name}{the name of the single frame loading function}
}
\value{
a list with two members: \code{name}, the name of the generated function and \code{code}, the code of the generated functions
}
\description{
This function invokes all the single code generators needed to build the loader code
}
