% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_result_data_frame_functions.R
\name{generate.result.data.frame.expand.function}
\alias{generate.result.data.frame.expand.function}
\title{Generate R Code for a Function which can Expand a Result Data Frame}
\usage{
generate.result.data.frame.expand.function(columns,
  function.name = .func.result.data.frame.expand,
  validate.function.name = .func.result.data.frame.validate)
}
\arguments{
\item{columns, }{the columns resulting from \code{create.columns} or
\code{join.columns}}

\item{function.name}{the name of the function}

\item{validate.function.name}{the name of the validation function}
}
\value{
a character array with the generated code
}
\description{
well, exactly that
}
