% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_result_data_frame_functions.R
\name{generate.result.data.frame.load.function}
\alias{generate.result.data.frame.load.function}
\title{Generate a Function that can Load a Result Data Frame from a CSV File}
\usage{
generate.result.data.frame.load.function(meta.columns, result.columns,
  function.name = .func.result.data.frame.load,
  validate.function.name = .func.result.data.frame.validate,
  expand.function.name = .func.result.data.frame.expand)
}
\arguments{
\item{meta.columns}{the meta-data columns. Their value can either be
specified as parameter of the generated function or inside the result csv
file.}

\item{result.columns}{the values of these columns can only be specified in
the csv file}

\item{function.name}{the function name to use for the generated function}

\item{validate.function.name}{the name of the function generated by
\code{generate.result.data.frame.validate.function}}

\item{expand.function.name}{the name of the function generated by
\code{generate.result.data.frame.expand.function}}
}
\value{
a function for loading a data frame from a csv file
}
\description{
Generate a loader function from a description of meta and result
  columns.
}
