% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_result_data_frame_functions.R
\name{generate.result.load.all.function}
\alias{generate.result.load.all.function}
\title{Generate Code to Load All Result Data and Create a Data Frame}
\usage{
generate.result.load.all.function(meta.columns, result.columns,
  function.name = .func.result.load.all,
  validate.function.name = .func.result.data.frame.validate,
  expand.function.name = .func.result.data.frame.expand,
  frame.load.function.name = .func.result.data.frame.load)
}
\arguments{
\item{meta.columns}{the meta-data columns. Their value can either be
specified as parameter of the generated function or inside the result csv
file.}

\item{result.columns}{the values of these columns can only be specified in
the csv file}

\item{function.name}{the function name to use for the generated function}

\item{validate.function.name}{the name of the function generated by
\code{generate.result.data.frame.validate.function}}

\item{expand.function.name}{the name of the function generated by
\code{generate.result.data.frame.expand.function}}

\item{frame.load.function.name}{the name of the single frame loading function}
}
\value{
a function for loading a data frame from a csv file
}
\description{
This function generate code to load all the single result
data frames and merges   them into one big one.
}
