% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_column.R
\name{join.columns}
\alias{join.columns}
\title{Merge Several Sets of Columns}
\usage{
join.columns(...)
}
\arguments{
\item{...}{the columns}
}
\value{
the merged column sets
}
\description{
Merge the sets stemming from \code{create.columns}
}
