% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_r_doc_references.R
\name{make.r.doc.references}
\alias{make.r.doc.references}
\title{Generate an R-Documentation Style References Section}
\usage{
make.r.doc.references(refs, bibliography, logger = NULL,
  subst.doi = "doi:\\\\\\\\href\\\\{\\\\2\\\\}\\\\{\\\\1\\\\}",
  subst.url = "\\\\\\\\url\\\\{\\\\1\\\\}",
  first.line.start = "#' @references ", first.line.end = "",
  first.line.as.separate.line = FALSE, make.text.before = NULL,
  make.text.after = NULL, normal.line.start = "#' ",
  normal.line.end = "", between.two.lines = "#'",
  after.first.line = "#'", sort = TRUE)
}
\arguments{
\item{refs}{an arbitrary collection of vectors with reference IDs}

\item{bibliography}{a bibliography data set}

\item{logger}{an optional logger function}

\item{subst.doi}{a substitute regular expression for dois: "\\1" is the doi,
"\\2" is the full link (optional, default value is supplied)}

\item{subst.url}{a substitute regular expression for urls, "\\1" is the url
(optional, default value is supplied)}

\item{first.line.start}{the starting string of the first line (optional,
default value is supplied)}

\item{first.line.end}{the ending string of the first line (optional, default
value is supplied)}

\item{first.line.as.separate.line}{should \code{first.line.start} be a
separate line (\code{TRUE}) or should it be the beginning pasted in front
of a normal line of references (\code{FALSE}) (optional, default value is
supplied)}

\item{make.text.before}{a function to which the reference id is passed and
which should return a string to be pasted before the actual reference text
(or \code{""}); \code{NULL} is supplied by default, meaning that no strings
are added}

\item{make.text.after}{a function to which the reference id is passed and
which should return a string to be pasted after the actual reference text,
(or \code{""}); \code{NULL} is supplied by default, meaning that}

\item{normal.line.start}{a string to paste at the beginning of every normal
line (optional, default value is supplied)}

\item{normal.line.end}{a string to paste at the end of every normal line
(optional, default value is supplied)}

\item{between.two.lines}{a string to be pasted as separate line between two
normal lines (optional, default value is supplied)}

\item{after.first.line}{the string to be inserted as line after the first
line (optional, default value is supplied)}

\item{sort}{should the bibliographic references be sorted?}
}
\value{
the multi-line reference text
}
\description{
Generate an R-Documentation style references section that can
  readily be inserted as an R documentation comment.
}
