% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_readme_toc.R
\name{make.readme.toc}
\alias{make.readme.toc}
\title{Number the Sections in a README.md File and Create a Table of Contents}
\usage{
make.readme.toc(text, min.level = 2L)
}
\arguments{
\item{text}{multiple lines of text of a README.md file}

\item{min.level}{the minimum section nesting level included in the table of
contents. This is set to 2, meaning that there should be at most one "#"
level with the overall title, while levels starting with "##" are included
into the table of contents.}
}
\value{
the text with sections numbered and a table of contents included
}
\description{
The sections in a README.md file are automatically numbered
  hierarchically. Then, the line containing the string "{toc}" is replaced
  with a table of contents.
}
