% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_references_text.R
\name{make.references.text}
\alias{make.references.text}
\title{Generate an References List}
\usage{
make.references.text(refs = NULL, bibliography, logger = NULL,
  subst.doi = "doi:\\\\\\\\href\\\\{\\\\2\\\\}\\\\{\\\\1\\\\}",
  subst.url = "\\\\\\\\url\\\\{\\\\1\\\\}",
  first.line.start = "#' @references ", first.line.end = "",
  first.line.as.separate.line = FALSE, make.text.before = NULL,
  make.text.after = NULL, normal.line.start = "#' ",
  normal.line.end = "", between.two.lines = "#'",
  after.first.line = between.two.lines)
}
\arguments{
\item{refs}{an arbitrary collection of vectors with reference IDs, if
\code{NULL}, then all bibliographic entries are used}

\item{bibliography}{a bibliography data set}

\item{logger}{an optional logger function}

\item{subst.doi}{a substitute regular expression for dois: "\\1" is the doi,
"\\2" is the full link}

\item{subst.url}{a substitute regular expression for urls, "\\1" is the url}

\item{first.line.start}{the starting string of the first line}

\item{first.line.end}{the ending string of the first line}

\item{first.line.as.separate.line}{should \code{first.line.start} be a
separate line (\code{TRUE}) or should it be the beginning pasted in front
of a normal line of references (\code{FALSE})}

\item{make.text.before}{a function to which the reference id is passed and
which should return a string to be pasted before the actual reference text
(or \code{""}); if \code{NULL} is supplied, no strings are added}

\item{make.text.after}{a function to which the reference id is passed and
which should return a string to be pasted after the actual reference text
(or \code{""}); if \code{NULL} is supplied, no strings are added}

\item{normal.line.start}{a string to paste at the beginning of every normal
line}

\item{normal.line.end}{a string to paste at the end of every normal line}

\item{between.two.lines}{a string to be pasted as separate line between two
normal lines}

\item{after.first.line}{the string to be inserted as line after the first
line}
}
\value{
the multi-line reference text
}
\description{
Generate an a textual list of references based on the given
  formatting instructions. This function is used by
  \link{make.r.doc.references} and \link{make.md.references}.
}
