% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_bibliography.R
\name{read.bibliography}
\alias{read.bibliography}
\title{Generate a Bibliographic List from a BibTeX File}
\usage{
read.bibliography(bib.file)
}
\arguments{
\item{bib.file}{the file}
}
\value{
a \code{data.frame} with four columns, namely \code{id}: the BibTeX
  entry's ID, \code{type}: the BibTeX entry's entry type, \code{bibtex}: the
  raw bibtex code, and \code{text} a formatted representation of the BibTeX
  that can be copy-pasted into package documentations
}
\description{
This function loads and parses a bibliography from the given
  bibtex file and turns it into a data frame. The goal is to allow for
  easy look-up and processing of BibTeX data along with an R package.
}
