% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{try.convert.numeric.to.int}
\alias{try.convert.numeric.to.int}
\title{Try to convert a numeric vector to an integer vector}
\usage{
try.convert.numeric.to.int(x, stopIfFails = FALSE, canFloor = FALSE)
}
\arguments{
\item{x}{the vector}

\item{stopIfFails}{stop if the conversion fails}

\item{canRound}{is it permitted to floor the integre values?}
}
\value{
the integer vector
}
\description{
Try to convert a numeric vector to an integer vector
}
