% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_md_references.R
\name{make.md.references}
\alias{make.md.references}
\title{Generate an Markdown Style References}
\usage{
make.md.references(refs, bibliography, first.line = "# References",
  logger = NULL)
}
\arguments{
\item{refs}{an arbitrary collection of vectors with reference IDs}

\item{bibliography}{a bibliography data set}

\item{first.line}{the first line}

\item{logger}{an optional logger function}
}
\value{
the multi-line reference text
}
\description{
Generate an Markdown formatted references.
}
