% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_r_doc_references.R
\name{make.r.doc.references}
\alias{make.r.doc.references}
\title{Generate an R-Documentation Style References Section}
\usage{
make.r.doc.references(refs, bibliography, logger = NULL)
}
\arguments{
\item{refs}{an arbitrary collection of vectors with reference IDs}

\item{bibliography}{a bibliography data set}

\item{logger}{an optional logger function}
}
\value{
the multi-line reference text
}
\description{
Generate an R-Documentation style references section that can
  readily be inserted as an R documentation comment.
}
